+++ app/src/main/java/woowacourse/shopping/data/CartRepository.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.data
+
+import woowacourse.shopping.domain.Product
+
+interface CartRepository {
+    fun cartProducts(): List<Product>
+
+    fun deleteCartProduct(id: Long)
+}
+
