+++ app/src/main/java/woowacourse/shopping/data/DefaultShoppingRepository.kt
@@ -45,23 +45,4 @@ class DefaultShoppingRepository : ShoppingRepository {
             "https://i.namu.wiki/i/YvceZuAFsjYzbrTKYS09muExzVUw0f5JFBTAOLeCJbyeKghRLpkDnc5_XmQ9KvOpyRqz3zSWVZq5DpeW0HToWQ.webp",
         )
     }
-
-    override fun cartProducts(): List<Product> {
-        return listOf(
-            Product(
-                3,
-                1000,
-                "오둥2",
-                "https://item.kakaocdn.net/do/8fb89536158119f901780df1ba184931a88f7b2cbb72be0bdfff91ad65b168ab",
-            ),
-            Product(
-                4,
-                1000,
-                "꼬상",
-                "https://w7.pngwing.com/pngs/921/264/png-transparent-chipmunk-chip-n-dale-sticker-the-walt-disney-company-goofy-others.png",
-            ),
-        )
-    }
-
-    override fun deleteCartProduct(id: Long) {}
 }
