+++ app/src/main/java/woowacourse/shopping/data/ShoppingRepository.kt
@@ -6,8 +6,5 @@ interface ShoppingRepository {
     fun products(amount: Int): List<Product>
 
     fun productById(id: Long): Product
-
-    fun cartProducts(): List<Product>
-
-    fun deleteCartProduct(id: Long)
 }
+
