+++ app/src/main/java/woowacourse/shopping/presentation/cart/ShoppingCartViewModel.kt
@@ -4,32 +4,32 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.data.ShoppingRepository
+import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.presentation.base.BaseViewModelFactory
 import woowacourse.shopping.presentation.shopping.product.ProductUi
 import woowacourse.shopping.presentation.shopping.product.toUiModel
 
 
 class ShoppingCartViewModel(
-    private val shoppingRepository: ShoppingRepository
+    private val cartRepository: CartRepository
 ) : ViewModel() {
     private val _products = MutableLiveData<List<ProductUi>>(emptyList())
     val products: LiveData<List<ProductUi>> get() = _products
 
     init {
-        _products.value = shoppingRepository.cartProducts().map { it.toUiModel() }
+        _products.value = cartRepository.cartProducts().map { it.toUiModel() }
     }
 
     fun deleteProduct(position: Int) {
         val deletedItem = _products.value?.get(position)
         if (deletedItem != null) {
             _products.value = _products.value?.minus(deletedItem)
-            shoppingRepository.deleteCartProduct(deletedItem.id)
+            cartRepository.deleteCartProduct(deletedItem.id)
         }
     }
 
     companion object {
-        fun factory(repository: ShoppingRepository): ViewModelProvider.Factory {
+        fun factory(repository: CartRepository): ViewModelProvider.Factory {
             return BaseViewModelFactory { ShoppingCartViewModel(repository) }
         }
     }
